Add files via upload
This is an android apk code called by me as Sauda app. In this app you can add item which you want to remember for example for making a list of items and once you bought it you can delete it also. It is a very user friendly apk and good project for beginners of android development. This is made using java in android studio.